+++ textrepo-app/src/test/java/nl/knaw/huc/resources/rest/VersionContentsResourceTest.java
@@ -4,12 +4,12 @@ import com.jayway.jsonpath.JsonPath;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import io.dropwizard.testing.junit5.ResourceExtension;
 import nl.knaw.huc.exceptions.MethodNotAllowedExceptionMapper;
+import nl.knaw.huc.helpers.ContentsHelper;
 import nl.knaw.huc.service.version.content.VersionContentsService;
 import org.glassfish.jersey.media.multipart.MultiPartFeature;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.io.IOException;
 import java.util.UUID;
 
 import static javax.ws.rs.client.Entity.entity;
@@ -20,16 +20,17 @@ import static org.mockito.Mockito.mock;
 @ExtendWith(DropwizardExtensionsSupport.class)
 public class VersionContentsResourceTest {
   private static final int CONTENT_DECOMPRESSION_LIMIT = 10;
+  private static final ContentsHelper CONTENTS_HELPER = new ContentsHelper(CONTENT_DECOMPRESSION_LIMIT);
 
   public static final ResourceExtension resource = ResourceExtension
       .builder()
       .addProvider(MultiPartFeature.class)
-      .addResource(new VersionContentsResource(mock(VersionContentsService.class), CONTENT_DECOMPRESSION_LIMIT))
+      .addResource(new VersionContentsResource(mock(VersionContentsService.class), CONTENTS_HELPER))
       .addResource(new MethodNotAllowedExceptionMapper())
       .build();
 
   @Test
-  public void testPost_returns405WithCorrectMsgAndAllowedMethods() throws IOException {
+  public void testPost_returns405WithCorrectMsgAndAllowedMethods() {
 
     var response = resource
         .client()