+++ CHANGELOG.md
@@ -14,6 +14,14 @@ Origin Link header to indexer fields request
 - Resources now send 'Content-Encoding: gzip' when sending gzip'd contents
 - Database queries using 'count' now use 'long' as their Java counterpart (was: int)
 
+## [v1.10]
+- [TT-631] When yielding contents, return gzip'd content when 'Accept-Encoding: gzip'
+- [TT-645] Resources now send 'Content-Encoding: gzip' when sending gzip'd contents
+- [TT-574] Database queries using 'count' now use 'long' as their Java counterpart (was: int)
+
+### Fixed
+- Use `tar.gz` instead of `tar` as backup extension
+
 ## [v1.9]
 ### Added
 - 'contentDecompressionLimit' (kB) configures plain/gzip output cutoff in HTTP results